// client/src/types/FormData.ts
export interface FormData  {
  _id?: string;
  fullname?: string;
  email?: string;
  password?: string;
  profileImage?: string;
  createdAt?: string;
};

export interface UserData {
  _id?: string;
  fullname?: string;
  email?: string;
  password?: string;
  profileImage?: string;
  createdAt?: string;
}